+++ crates/primitives/src/lib.rs
@@ -13,7 +13,6 @@ pub use self::chain_type::ChainType;
 pub mod chain_evm;
 pub use self::chain_evm::EVMChain;
 mod macros;
-mod evm_chain_config;
 pub mod chain_bitcoin;
 pub use self::chain_bitcoin::BitcoinChain;
 pub mod name;
